Move `matches` lint declarations into sub-modules
