+++ clippy_lints_matches/src/declared_lints.rs
@@ -3,33 +3,33 @@
 // Manual edits will be overwritten.
 
 pub static LINTS: &[&::declare_clippy_lint::LintInfo] = &[
-    crate::COLLAPSIBLE_MATCH_INFO,
-    crate::INFALLIBLE_DESTRUCTURING_MATCH_INFO,
-    crate::MANUAL_FILTER_INFO,
-    crate::MANUAL_MAP_INFO,
-    crate::MANUAL_OK_ERR_INFO,
-    crate::MANUAL_UNWRAP_OR_INFO,
-    crate::MANUAL_UNWRAP_OR_DEFAULT_INFO,
-    crate::MATCH_AS_REF_INFO,
-    crate::MATCH_BOOL_INFO,
-    crate::MATCH_LIKE_MATCHES_MACRO_INFO,
-    crate::MATCH_OVERLAPPING_ARM_INFO,
-    crate::MATCH_REF_PATS_INFO,
-    crate::MATCH_SAME_ARMS_INFO,
-    crate::MATCH_SINGLE_BINDING_INFO,
-    crate::MATCH_STR_CASE_MISMATCH_INFO,
-    crate::MATCH_WILD_ERR_ARM_INFO,
-    crate::MATCH_WILDCARD_FOR_SINGLE_VARIANTS_INFO,
-    crate::NEEDLESS_MATCH_INFO,
-    crate::REDUNDANT_GUARDS_INFO,
-    crate::REDUNDANT_PATTERN_MATCHING_INFO,
-    crate::REST_PAT_IN_FULLY_BOUND_STRUCTS_INFO,
-    crate::SIGNIFICANT_DROP_IN_SCRUTINEE_INFO,
-    crate::SINGLE_MATCH_INFO,
-    crate::SINGLE_MATCH_ELSE_INFO,
-    crate::TRY_ERR_INFO,
-    crate::WILDCARD_ENUM_MATCH_ARM_INFO,
-    crate::WILDCARD_IN_OR_PATTERNS_INFO,
+    crate::collapsible_match::COLLAPSIBLE_MATCH_INFO,
+    crate::infallible_destructuring_match::INFALLIBLE_DESTRUCTURING_MATCH_INFO,
+    crate::manual_filter::MANUAL_FILTER_INFO,
+    crate::manual_map::MANUAL_MAP_INFO,
+    crate::manual_ok_err::MANUAL_OK_ERR_INFO,
+    crate::manual_unwrap_or::MANUAL_UNWRAP_OR_INFO,
+    crate::manual_unwrap_or::MANUAL_UNWRAP_OR_DEFAULT_INFO,
     crate::map_unit_fn::OPTION_MAP_UNIT_FN_INFO,
     crate::map_unit_fn::RESULT_MAP_UNIT_FN_INFO,
+    crate::match_as_ref::MATCH_AS_REF_INFO,
+    crate::match_bool::MATCH_BOOL_INFO,
+    crate::match_like_matches::MATCH_LIKE_MATCHES_MACRO_INFO,
+    crate::match_ref_pats::MATCH_REF_PATS_INFO,
+    crate::match_same_arms::MATCH_SAME_ARMS_INFO,
+    crate::match_single_binding::MATCH_SINGLE_BINDING_INFO,
+    crate::match_str_case_mismatch::MATCH_STR_CASE_MISMATCH_INFO,
+    crate::match_wild_enum::MATCH_WILDCARD_FOR_SINGLE_VARIANTS_INFO,
+    crate::match_wild_enum::WILDCARD_ENUM_MATCH_ARM_INFO,
+    crate::match_wild_err_arm::MATCH_WILD_ERR_ARM_INFO,
+    crate::needless_match::NEEDLESS_MATCH_INFO,
+    crate::overlapping_arms::MATCH_OVERLAPPING_ARM_INFO,
+    crate::redundant_guards::REDUNDANT_GUARDS_INFO,
+    crate::redundant_pattern_match::REDUNDANT_PATTERN_MATCHING_INFO,
+    crate::rest_pat_in_fully_bound_struct::REST_PAT_IN_FULLY_BOUND_STRUCTS_INFO,
+    crate::significant_drop_in_scrutinee::SIGNIFICANT_DROP_IN_SCRUTINEE_INFO,
+    crate::single_match::SINGLE_MATCH_INFO,
+    crate::single_match::SINGLE_MATCH_ELSE_INFO,
+    crate::try_err::TRY_ERR_INFO,
+    crate::wild_in_or_pats::WILDCARD_IN_OR_PATTERNS_INFO,
 ];
