+++ clippy_lints_matches/src/manual_utils.rs
@@ -1,5 +1,5 @@
-use super::MATCH_AS_REF;
 use crate::map_unit_fn::OPTION_MAP_UNIT_FN;
+use crate::match_as_ref::MATCH_AS_REF;
 use clippy_utils::source::{snippet_with_applicability, snippet_with_context};
 use clippy_utils::sugg::Sugg;
 use clippy_utils::ty::{is_copy, is_type_diagnostic_item, peel_mid_ty_refs_is_mutable, type_is_unsafe_function};
