+++ clippy_lints_matches/src/wild_in_or_pats.rs
@@ -4,7 +4,36 @@ use rustc_hir::{Arm, Expr, PatKind};
 use rustc_lint::LateContext;
 use rustc_middle::ty;
 
-use super::WILDCARD_IN_OR_PATTERNS;
+declare_clippy_lint! {
+    /// ### What it does
+    /// Checks for wildcard pattern used with others patterns in same match arm.
+    ///
+    /// ### Why is this bad?
+    /// Wildcard pattern already covers any other pattern as it will match anyway.
+    /// It makes the code less readable, especially to spot wildcard pattern use in match arm.
+    ///
+    /// ### Example
+    /// ```no_run
+    /// # let s = "foo";
+    /// match s {
+    ///     "a" => {},
+    ///     "bar" | _ => {},
+    /// }
+    /// ```
+    ///
+    /// Use instead:
+    /// ```no_run
+    /// # let s = "foo";
+    /// match s {
+    ///     "a" => {},
+    ///     _ => {},
+    /// }
+    /// ```
+    #[clippy::version = "1.42.0"]
+    pub WILDCARD_IN_OR_PATTERNS,
+    complexity,
+    "a wildcard pattern used with others patterns in same match arm"
+}
 
 pub(crate) fn check(cx: &LateContext<'_>, expr: &Expr<'_>, arms: &[Arm<'_>]) {
     // first check if we are matching on an enum that has the non_exhaustive attribute
